feat(aio): add Slack destination to scout creation
