+++ frontend/src/scenes/inbox/components/config/scouts/ScoutCreateModal.tsx
@@ -21,6 +21,7 @@ import {
     SCOUT_DAILY_AT_SCHEDULE_MODE,
     SIGNALS_SCOUT_SKILL_PREFIX,
 } from '../../../utils/scoutRunsWindow'
+import { ScoutSlackDestination } from './ScoutSlackDestination'
 
 export interface ScoutCreateModalProps {
     isOpen: boolean
@@ -186,6 +187,15 @@ export function ScoutCreateModal({ isOpen, onClose, initialValues, onCreated }: 
                         <span className="text-xs text-muted">
                             Turn off inbox findings to run the scout in dry-run mode.
                         </span>
+                        <LemonField name="config.output_destinations">
+                            {({ value, onChange }) => (
+                                <ScoutSlackDestination
+                                    destination={value?.slack}
+                                    onChange={onChange}
+                                    disabledReason={isScoutCreateFormSubmitting ? 'Creating the scout' : undefined}
+                                />
+                            )}
+                        </LemonField>
                     </div>
                 </div>
             </Form>
