+++ frontend/src/scenes/inbox/logics/scoutCreateModalLogic.test.ts
@@ -70,6 +70,12 @@ describe('scoutCreateModalLogic', () => {
                     enabled: false,
                     emit: false,
                     run_interval_minutes: 60,
+                    output_destinations: {
+                        slack: {
+                            integration_id: 42,
+                            channel: 'C123|#ai-observability',
+                        },
+                    },
                 },
             },
             onClose,
@@ -87,6 +93,12 @@ describe('scoutCreateModalLogic', () => {
                 emit: false,
                 run_interval_minutes: 60,
                 run_cron_schedule: null,
+                output_destinations: {
+                    slack: {
+                        integration_id: 42,
+                        channel: 'C123|#ai-observability',
+                    },
+                },
             },
         })
 
@@ -101,6 +113,12 @@ describe('scoutCreateModalLogic', () => {
                 emit: false,
                 run_interval_minutes: 60,
                 run_cron_schedule: null,
+                output_destinations: {
+                    slack: {
+                        integration_id: 42,
+                        channel: 'C123|#ai-observability',
+                    },
+                },
             },
         })
         expect(onCreated).toHaveBeenCalledWith(CREATED_SCOUT)
